Create downsampleZ box and listener
This commit begins the changes necessary to create a separate option for
downsampling in Z, if preserve anisotropy is chosen.

This series of changes is intended for stitched microscopy datasets,
which usually have different resolutions in Z and XY.  At times, it is
valuable to downsample in XY and not in Z.  Once fully implemented,
this change will let you easily select this as an option.